Fix Executor result

+++ simulator/src/Executor.cpp
@@ -1,6 +1,7 @@
 #include "../include/Executor.hpp"
 #include "../include/ALU.hpp"
 #include "../include/CPU.hpp"
+#include <iostream>
 
 void Executor::executeRType(const Instruction& instr, CPU& cpu) {
     int32_t rs1 = static_cast<int32_t>(cpu.readReg(instr.rs1));
@@ -10,26 +11,26 @@ void Executor::executeRType(const Instruction& instr, CPU& cpu) {
 
     switch (instr.funct3) {
 
-        case 0b000: // ADD/SUB
-            if (instr.funct7 == 0b0100000)
-                result = ALU::subSigned(rs1, rs2);
-            else
-                result = ALU::addSigned(rs1, rs2);
+        case 0b000: // ADD / SUB
+            if (instr.funct7 == 0b0100000) // SUB
+                result = static_cast<uint32_t>(ALU::subSigned(rs1, rs2));
+            else // ADD
+                result = static_cast<uint32_t>(ALU::addSigned(rs1, rs2));
             break;
 
         case 0b111: // AND
-            result = ALU::andOp(rs1, rs2);
+            result = ALU::andOp(static_cast<uint32_t>(rs1), static_cast<uint32_t>(rs2));
             break;
 
         case 0b110: // OR
-            result = ALU::orOp(rs1, rs2);
+            result = ALU::orOp(static_cast<uint32_t>(rs1), static_cast<uint32_t>(rs2));
             break;
 
         case 0b100: // XOR
-            result = ALU::xorOp(rs1, rs2);
+            result = ALU::xorOp(static_cast<uint32_t>(rs1), static_cast<uint32_t>(rs2));
             break;
 
-        case 0b010: // SLT
+        case 0b010: // SLT signed
             result = ALU::sltSigned(rs1, rs2);
             break;
 
@@ -44,14 +45,14 @@ void Executor::executeRType(const Instruction& instr, CPU& cpu) {
 
 void Executor::executeIType(const Instruction& instr, CPU& cpu) {
     int32_t rs1 = static_cast<int32_t>(cpu.readReg(instr.rs1));
-    int32_t imm = instr.imm;  
+    int32_t imm = instr.imm;  // já sign-extended pelo decoder
 
     uint32_t result = 0;
 
     switch (instr.funct3) {
 
         case 0b000: // ADDI
-            result = ALU::addSigned(rs1, imm);
+            result = static_cast<uint32_t>(ALU::addSigned(rs1, imm));
             break;
 
         case 0b100: // XORI
@@ -66,7 +67,7 @@ void Executor::executeIType(const Instruction& instr, CPU& cpu) {
             result = ALU::andOp(static_cast<uint32_t>(rs1), static_cast<uint32_t>(imm));
             break;
 
-        case 0b010: // SLTI
+        case 0b010: // SLTI signed
             result = ALU::sltSigned(rs1, imm);
             break;
 
@@ -77,4 +78,3 @@ void Executor::executeIType(const Instruction& instr, CPU& cpu) {
 
     cpu.writeReg(instr.rd, result);
 }
-
